+++ packages/core/src/index.ts
@@ -1,4 +1,6 @@
 export * from './base58.js';
+export * from './document-syntax.js';
+export * from './container-syntax.js';
 export * from './front-matter.js';
 export * from './hash.js';
 export * from './id.js';